tweaked routes; still having troubele
