+++ controllers/api/index.js
@@ -1,9 +1,9 @@
 const router = require('express').Router();
-const usersRoutes = require('./userRoutes')
+const userRoutes = require('./userRoutes')
 const animalRoutes = require('./animalRoutes')
 
 // all of these routes are prefixed with '/api'
-router.use('/', usersRoutes);
+router.use('/users', userRoutes);
 router.use('/animals', animalRoutes);
 
 module.exports = router;
