+++ controllers/homeRoutes.js
@@ -1,0 +1,7 @@
+const router = require('express').Router();
+
+router.get('/', (req, res) => {
+    res.render('home');
+});
+
+module.exports = router;
