+++ controllers/index.js
@@ -1,8 +1,8 @@
 const router = require('express').Router();
 const apiRoutes = require('./api');
-const htmlRoutes = require('./htmlRoutes')
+const homeRoutes = require('./homeRoutes')
 
 router.use('/api', apiRoutes);
-router.use('/', htmlRoutes)
+router.use('/', homeRoutes)
 
 module.exports = router;
