+++ server.js
@@ -5,6 +5,8 @@ const session = require('express-session');
 const exphbs = require('express-handlebars');
 
 const routes = require('./controllers');
+// const apiRoutes = require('./controllers/api');
+
 // const helpers = require('./utils/helpers');
 
 const passport = require('passport');
@@ -63,12 +65,12 @@ passport.use(new LocalStrategy(function verify(username, password, cb) {
 */
 app.engine('handlebars', hbs.engine);
 app.set('view engine', 'handlebars');
-// app.set('views', path.join(__dirname, './views'));
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use(routes);
+// app.use(apiRoutes);
 
 sequelize.sync({ force: false }).then(() => {
   app.listen(PORT, () => console.log('Now listening'));
