+++ views/home.handlebars
@@ -1,7 +1,7 @@
 <div class='container login inter-tight-regular'>
   <div class='login_form'>
     <h2>Login</h2>
-    <form id='login-form' action='/api/login/password' method='POST'>
+    <form id='login-form' action='/api/users/login/password' method='POST'>
       <div>
         <label for='login-email'>Email:</label>
         <input type='text' id='login-email' name='username' required>
@@ -16,7 +16,7 @@
   </div>
   <div class='register_form'>
       <h2>Register</h2>
-      <form id='register-form' action='api/register' method='POST'>
+      <form id='register-form' action='api/users/register' method='POST'>
         <div>
           <label for='register-user'>Username:</label>
           <input type='text' id='register-user' name='user' required>
